feat: move authpublickey to proto
