+++ deployer/Cargo.toml
@@ -42,3 +42,6 @@ uuid = { workspace = true, features = ["v4"] }
 [dependencies.shuttle-common]
 workspace = true
 features = ["backend", "models", "openapi"]
+
+[dependencies.shuttle-proto]
+workspace = true
