+++ deployer/prepare.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env sh
+
+###############################################################################
+# This file is used by our common Containerfile incase the container for this #
+# service might need some extra preparation steps for its final image         #
+###############################################################################
+
+# Nothing to prepare in container image here
