+++ proto/auth.proto
@@ -61,8 +61,6 @@ message PublicKeyResponse {
 
 message PublicKeyRequest {}
 
-message Request {}
-
 message ConvertCookieRequest {}
 
 message LogoutRequest {}
